requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=8.3.0
opencv-python>=4.5.0
ptwt>=0.1.0
timm>=0.6.0
scipy>=1.7.0
pandas>=1.3.0 